Alert message disappears and reset the form fields
I introduce myself. I have seen your video on YouTube (I made a comment on the video, my channel name is "Chitorra Garcia").

I saw your code and I found that the "reset ()" method can be very well liked for when the form fields are reset when the sending is successful and the fields are not filled.
And... the other thing is the alert message.

I hope you can accept my pull resquest. Excellent video and code.
Greetings!